requirements.txt
asana
openai
requests
python-dotenv
